pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scalp"
version = "0.1.0"
description = "Delta-neutral gamma scalping engine for OKX linear derivatives"
# Target: Python 3.14t (free-threaded / PEP 703 no-GIL build).
# Minimum is 3.13 where the free-threaded build first shipped as stable.
requires-python = ">=3.13"
dependencies = [
    "python-okx>=0.3.9",
    "trio>=0.27",
    "trio-websocket>=0.11",
    "numpy>=2.0",
    "scipy>=1.14",
    "pydantic-settings>=2.5",
]

[project.scripts]
scalp = "scalp.main:run"

[tool.hatch.build.targets.wheel]
packages = ["scalp"]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "standard"
